requirements.txt
numpy==1.26.4
torch==2.2.0
huggingface_hub<0.20
pyannote.audio==3.3.2
soundfile==0.13.1
groq==0.31.0
pydub==0.25.1
python-dotenv==1.0.0
fastapi==0.116.2
uvicorn==0.22.0
pymongo
python-multipart
dotenv
scikit-learn
